Added report from 2017
This report was the culmination of a semester worth of assignments for 'Introduction to Data Analytics' in 2017. 

The useful components of this report are:
 - descriptions of how the data was obtained
 - description of data cleansing processes
 - exploration of data properties

This report also describes some models that I generated at the time, and their performance metrics. However, I do not plan on reusing these models, only the data.